Add register Link

+++ Web/routes/api/user.php
@@ -15,13 +15,11 @@ Route::group(['prefix' => 'user', 'middleware' => ['api']], function () {
 
     Route::post('/', 'UserController@store');
 
-    Route::post('/login', 'UserController@login');
+    Route::get('/{id}', 'UserController@show')->where('id', '[0-9]+')->middleware('auth');
 
-    Route::get('/{id}', 'UserController@show')->where('id', '[0-9]+');
+    Route::patch('/{id}', 'UserController@update')->where('id', '[0-9]+')->middleware('auth');
 
-    Route::patch('/{id}', 'UserController@update')->where('id', '[0-9]+');
-
-    Route::delete('/{id}', 'UserController@destroy')->where('id', '[0-9]+');
+    Route::delete('/{id}', 'UserController@destroy')->where('id', '[0-9]+')->middleware('auth');
 
     Route::get('/search/{text}', 'UserController@search');
